requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.1.0
sqlparse==0.5.3

djangorestframework~=3.16.0